crud correct work

+++ src/main/java/ru/javawebinar/topjava/web/MealServlet.java
@@ -29,36 +29,51 @@ public class MealServlet extends HttpServlet {
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
-        if (request.getParameter("id") != null) {
+        if (request.getParameter("del") != null && request.getParameter("id") != null) {
             String id = request.getParameter("id");
-            if (request.getParameter("del") != null)
-                mealDao.delete(id);
+            mealDao.delete(id);
         }
-
-        if (request.getParameter("add") != null) {
-            String description = request.getParameter("description");
-            int calories = Integer.parseInt(request.getParameter("calories"));
-            LocalDateTime localDateTime = LocalDateTime.parse(request.getParameter("date"), DateTimeFormatter.ofPattern(PATTERN));
-            mealDao.add(new Meal(localDateTime, description, calories));
+        if ("add".equals(request.getParameter("formtype"))) {
+            Meal meal = createMealByRequest(request);
+            mealDao.add(meal);
         }
-        /*if (Boolean.parseBoolean(request.getParameter("edit"))) {
-            Meal meal = mealDao.getById(request.getParameter("description"));
-            String description = request.getParameter("description");
-            int calories = Integer.parseInt(request.getParameter("calories"));
-            LocalDateTime localDateTime = null;//LocalDateTime.parse(request.getParameter("date"), DateTimeFormatter.ofPattern(PATTERN));
-            meal.setDescription(description);
-            meal.setCalories(calories);
-            meal.setDateTime(localDateTime);
+        if ("edit".equals(request.getParameter("formtype")) && request.getParameter("id") != null) {
+            Meal meal = createMealByRequest(request);
             mealDao.update(meal);
-        }*/
-        /*add = true;
-        else add = false;*/
-
+        }
         List<MealWithExceed> mealsWithExceed = MealsUtil.getFilteredWithExceeded(new ArrayList<>(MealsUtil.MEALS_BASE.values()), 2000);
         request.setAttribute("mealsWithExceedList", mealsWithExceed);
 
         RequestDispatcher requestDispatcher = request.getRequestDispatcher("/meals-jsp");
         requestDispatcher.forward(request, response);
+    }
+
+    private Meal createMealByRequest(HttpServletRequest request) {
+        Meal meal=new Meal();
+        if (request.getParameter("id") != null) {
+            meal.setId(request.getParameter("id"));
+        }
+        String description;
+        try {
+            description = request.getParameter("description");
+        } catch (Exception e) {
+            description = "no description";
+        }
+        int calories;
+        try {
+            calories = Integer.parseInt(request.getParameter("calories"));
+        } catch (Exception e) {
+            calories = 0;
+        }
+        LocalDateTime localDateTime;
+        try {
+            localDateTime = LocalDateTime.parse(request.getParameter("date"), DateTimeFormatter.ofPattern(PATTERN));
+        } catch (Exception e) {
+            localDateTime = LocalDateTime.now();
+        }
+        meal.setDescription(description);
+        meal.setCalories(calories);
+        meal.setDateTime(localDateTime);
+        return meal;
     }
 }
